Refatorar classes DFS e BFS para estenderem super classe XFS
Evitar código duplicado
